EY provider remove submission controller
