+++ app/forms/journeys/early_years_payment/provider/authenticated/confirmation_form.rb
@@ -1,0 +1,10 @@
+module Journeys
+  module EarlyYearsPayment
+    module Provider
+      module Authenticated
+        class ConfirmationForm < Form
+        end
+      end
+    end
+  end
+end
