+++ app/models/journeys/early_years_payment/provider/authenticated/slug_sequence.rb
@@ -19,6 +19,7 @@ module Journeys
 
           RESULTS_SLUGS = %w[
             check-your-answers
+            confirmation
             ineligible
           ].freeze
