+++ app/views/early_years_payment/provider/authenticated/claims/confirmation.html.erb
@@ -45,4 +45,3 @@
     </p>
   </div>
 </div>
-
